gsc-ai-backend/requirements.txt
# MCP server framework — official Anthropic Python SDK
mcp[cli]>=1.0.0

# Environment variable loading from .env file
python-dotenv>=1.0.0

# AI Orchestrator dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.0.0
pydantic>=2.0.0
websockets>=12.0
pytest>=7.0.0
